ref: move sentry.{features,options}.rollout
resolves an import cycle between features<=>options and the rollout module has nothing to do with sentry.features
