+++ pyproject.toml
@@ -432,6 +432,7 @@ module = [
     "sentry.nodestore.filesystem.backend",
     "sentry.nodestore.models",
     "sentry.notifications.services.*",
+    "sentry.options.rollout",
     "sentry.organizations.*",
     "sentry.ownership.*",
     "sentry.plugins.base.response",
